Changed int to long

+++ src/main/java/ntnu/idatt1002/controllers/TaskController.java
@@ -18,7 +18,7 @@ import java.io.IOException;
 
 public class TaskController {
 
-    private int taskId;
+    private long taskId;
     @FXML private Text taskName;
     @FXML private Text taskDescription;
     @FXML private Label taskDate;
@@ -128,7 +128,7 @@ public class TaskController {
 
     }
 
-    public void setTaskId(int id){
+    public void setTaskId(long id){
         this.taskId = id;
     }
 }
